log_cgorup_info and log_cgroup_warning should be in cgroupapi
