+++ azurelinuxagent/ga/cgroupstelemetry.py
@@ -17,21 +17,8 @@ import errno
 import threading
 
 from azurelinuxagent.common import logger
-from azurelinuxagent.common.event import WALAEventOperation, add_event
 from azurelinuxagent.ga.cgroup import CpuCgroup
 from azurelinuxagent.common.future import ustr
-
-
-def log_cgroup_info(formatted_string, op=WALAEventOperation.CGroupsInfo, send_event=True):
-    logger.info("[CGI] " + formatted_string)
-    if send_event:
-        add_event(op=op, message=formatted_string)
-
-
-def log_cgroup_warning(formatted_string, op=WALAEventOperation.CGroupsInfo, send_event=True):
-    logger.info("[CGW] " + formatted_string)  # log as INFO for now, in the future it should be logged as WARNING
-    if send_event:
-        add_event(op=op, message=formatted_string, is_success=False, log_event=False)
 
 
 class CGroupsTelemetry(object):
